Permissions: Avoid RestrictionStore lookup for unsupported actions
When rendering Special:RecentChanges for an user with rollback rights,
the software checks whether the user has permission to perform the
rollback action on each entry via User::probablyCan(), which
delegates to PermissionManager. This in turn checks for each entry
whether the title has protections (e.g. from action=protect) against the
"rollback" action, which is unlikely to be the case since by default
wgRestrictionTypes only supports create/edit/move/upload.

This triggers noticeable memcached I/O and overhead.[1]

So, as an optimization, avoid fetching page restrictions in
RestrictionStore for action types that are known to not be restrictable
via page protection. In the unlikely event that wgRestrictionTypes is
customised to offer separate restrictions for "rollback", this will
work continue to work as expected and is covered by tests.

-
[1] https://gist.github.com/mszabo-wikia/ba0af0af63cde7dd9cc289672a58d923/raw/acd6281072d5dc1cd58b7319a6e69b0e6a75d7c5/recentchanges.speedscope.json

Bug: T341319
Change-Id: I912e4ef2fa3d48238c01631d4f35940aea93ab03