phpunit: Delete DummyExtensionsTest
A code comment says that this code is necessary to avoid warnings like
"No tests in class ...", but I wasn't able to reproduce the warning in
question.

This code was added in https://w.wiki/73Wv, in 2010. I guess PHPUnit
might have changed enough over the last 13 years to no longer emit this
warning.

A quick look at the TestSuite class suggests that the warning is not
emitted when the constructor is called with a string, which is what
we're doing.

Things might change in PHPUnit 10, but TestSuite::__construct was made
private, so I think we may have to rewrite our suites anyway.

Change-Id: Ie044900ffcffca9c8f0d96b425271a2eda90a5cd