phpunit: Avoid hardcoded URL in UserRightsProxyTest
The hardcoded value doesn't work in all MW installs, because it depends
on the configured URL scheme. For instance, the test fails with the
default MediaWiki-Docker setup.

This test actually doesn't need URLs at all, so avoid unnecessary URL
conversion and config mocks, and just assert that the page title is the
one we expect.

Change-Id: If7ed01f7c2404bbd8a2751131647960816828949