phpunit: Do not setup the test DB for tests that don't need it
If a test needs to do something with the database, it should be in the
`Database` group. Therefore, there should be no need to setup the DB if
that isn't the case.

Also make sure that the real DB cannot be accessed if the test is not in
the Database group.

MediaWikiIntegrationTestCaseNoDbTest was partly copied from I2b2eff78.

Bug: T155147
Depends-On: I11824a635fb43cfac8560b8fdecd556c82b9ce42
Depends-On: Ia12e28b4e23f111c0b4288e9ac19d3ad10f30d6c
Change-Id: I96ecf9ffd5fbf18db3e15e8774f7dc2fe5b9fd2a