phpunit: Make getTestUser and friends non-static
Making these methods static was a mistake. They really shouldn't be
static, because they interact with the database, and should only be
called after initialization.

The goal is to add a needsDB() check to them, but this will be done in
another patch. Note that TestUser::assertNotReal() already checks
that the user is not in the real database, but it doesn't warn the test
author that their (potentially non-database) test is using the test
database.

Also make them protected, like the other helper methods. If someone
needs to use these outside of MediaWikiIntegrationTestCase, provided
that their use case is sensible, they can use TestUserRegistry directly.

This change should mostly work as long as getTestUser etc. are only
called in non-static methods, regardless of whether callers call them
statically. But if these methods are called in a static method it might
be a bad sign (e.g., if it's a data provider).

For historical reasons, many places already call these methods
non-statically, which is a good thing. A quick codesearch suggests that
around 464 files call them non-statically, and 181 call them statically.

Change-Id: I04dfa8a1a4831767b90bc5a30c4a706069670c9d