phpunit: Do not call addCoreDBData if the test doesn't need the DB
This is a small step towards T155147. For now we clone the database even
if the test doesn't say it needs it, but we stop adding core data.

If a non-database test depends on this data, it means that it really
should be in the database group, or that it's otherwise doing something
wrong. For example, both MediaWikiTest and
RefreshSecondaryDataUpdateTest were depending on the 'UTPage' page being
created by the test framework, although they did that in such a way that
the needsDB() check in getExistingTestPage() weren't triggered.

Bug: T155147
Depends-On: I1a94c6ca2f335ee5a2d7d57df6dc46b65ca1f767
Change-Id: Ia04ff528628a7ae8b350e3122f5a06dbc1ff753b