UserGroupManagerTest: clear all hooks before running tests
There are several user groups-related hooks that can interfere with this
test. For instance, if AbuseFilter is installed, some tests will fail
because AF hooks call some methods on a no-op User mock. This class is
already making sure that the needed hooks it wants to test are set, so
there's no need to run any other hooks.

Change-Id: Iff184b5de1b150fd8043af13f16f8e1c870eb3fe