MediaWikiTest: avoid leaking values into the global state
The MediaWiki class is the quintessence of global state. I'm not even
sure if it's safe to run this test together with others. At any rate,
avoid using RequestContext::getMain(), given that the context is passed
in explicitly. And also reset global $wgTitle after the test, or any
value set in this test may persist forever.

Bug: T341951
Change-Id: I1a94c6ca2f335ee5a2d7d57df6dc46b65ca1f767