Throw exception in getTestUser etc. if the test doesn't need the DB
These methods create the test users, meaning they need a database, and
shouldn't be used if the test is not in the database group.

Note that TestUser::assertNotReal() already checks that the user is
not in the real database, but it doesn't warn the test author that
their (potentially non-database) test is using the test database.

Bug: T155147
Depends-On: I211cb60296e5c2446128fcdf2caaadc728a8c272
Change-Id: Ic4a72fbfaee730b8417848ae0603443d4995fefc