ApiTestCase: Clone User object when Authority is used
Something in Session(Backend? Manager?) replaces the user with an
unknown user if it doesn't exist, which later causes a token mismatch.

I didn't find the root cause for this issue, as I stopped digging once I
found a working solution, but it shouldn't realy matter as this is only
test code and we're moving away from User anyway.

Bug: T341953
Change-Id: I6b675a8b1c6b3d7fff8ca1330f8857b7393ab012