phpunit: Set the MW_INSTALL_PATH env variable if needed
phpunit.php did that, and also the bootstrap file before
Ie20617ac80d12f99ceaeae53b852cc6ed76b3feb. I was under the impression
that wfDetectInstallPath would also set this variable, but it doesn't --
it only sets $IP and the MW_INSTALL_PATH constant.

So restore the bit of code that sets the env variable, but do that only
in the bootstrap, and not in wfDetectInstallPath, because it's
presumably not needed there.

Change-Id: Icbd718bdea3567fb85d39926bed93af72de60c9b