objectcache: Remove unused WRAPPER_COLLECTION_CALLBACK feature
This was introduced for use by WANObjectCache, however, this never
worked in practice.

The feature only influences determineKeyPrefixForStats(), which in
turn is only called by MediumSpecificBagOStuff::updateOpStats, which
in turn is exclusively called within BagOStuff subclasses that aren't
used with WANObjectCache (RESTBag, RedisBag, SqlBag).

The next commit simplifies BagOStuff::makeKey, which can make a bigger
impact if this feature doesn't need to be supported.

During drafting of the next commit, I found that the feature doesn't
appear to work well in practice, because getCollectionFromSisterKey
did not account for mcrouter-style route prefixes, as inserted by
makeSisterKey().

Change-Id: Iecea959c88abbd7a0a17b92c1c4d854602748280