TempUserCreator: Change getStashedNameOrPlaceholder() to just getStashedName()
I found myself needing to just get the stashes name without the
placeholder for T331397, and I thought that this seems like a weird
API. The caller should get the placeholder separately if desired,
which is easy to do in all of the current callers.

Follow-up to 65f04c763aa062d90a4c72cab222bb068a433aa0.

Change-Id: Icf1d24f17b60a461ef682bc5da01be6ebc3a2d93