Restore DifferenceEngine::localiseLineNumbers()
Since it’s needed by EntitySchema, and because it was also, as far as I
can tell, removed without following the stable interface policy in
change I5c506e3916 (commit 2aa87cdf2c).

Not marked as soft-deprecated because there doesn’t seem to be a way to
access the replacement.

Bug: T342209
Change-Id: Iba9fd8bb56ea0e1543c9a9beab85b01f9549fbd7