Add ResourcesTest to the Database group
This tests needs the database in various places:
SqlModuleDependencyStore, wfGetDB() for WikiModule::preloadTitleInfo,
and other things in WikiModule. Some of these might be avoided; for
instance, it could use KeyValueDependencyStore. Others can't really be
avoided as the relevant classes call wfGetDB (or equivalent methods in
MediaWikiServices) directly.

Given the importance of this test, I'd rather have it continue using the
database, so add it to the Database group.

Change-Id: I564a826e97967897a3538d74cdef6f3dbdd570f1