Add SlotDiffRenderer::localizeDiff()
It makes sense for SlotDiffRenderer to localize the diffs that it
returns.

DifferenceEngine can't conveniently use this right now since it stores
all slot diffs concatenated together in a cache entry. It can't break
the string apart again to call the right SlotDiffRenderer on each
slot diff. So it assumes there must be a text diff in there somewhere
and calls TextDiffer over the whole thing.

EntitySchema does not have this problem and can benefit from this
method.

Bug: T342209
Change-Id: Ie06e100bb0e945de5dbe3dade14b2e6b706e078d