Improve collapsing on the enhanced changes list
The newline after the <input> adds a wrapping <p> on an include by
  {{Special:RecentChanges|enhanced=1}}
Prevent this by put the newline into a comment.

The hack with tr:not( :first-of-type ) is not needed, because there is
the class mw-rcfilters-ui-highlights-enhanced-nested for this case.

This is a follow-up to 9bf98ab9dbdd50eca76dff92a5f896e49e97e687.

Change-Id: I53527548b49d1951490414b692edaef65df88d48