Improve performance of LocalisationCache::mergeMagicWords()
While this is not a critical bottleck, I have seen it popping
up in some recent flamegraphs (see T225730). This is part of a
multi-dimensional loop and can be executed a lot during tests.

This rewrite makes this method about twice as fast.

A key change is that we don't need to strip the array keys with
array_values(). This happens anyway.

Bug: T225730
Change-Id: Iae7a18d4848910b53fa2868dcbb5d15ed88bc659