LocalisationCache: Reset loaded(Sub)Items when overwriting data
If we’re replacing all the data for a given language code, also reset
the loaded items and sub-items for that language code (before marking
some items as loaded again). This seemed to be required for some tests
to pass in change I7ec2d87c0f.

In initLanguage(), we can avoid this by instead merging the preload data
into the existing data, rather than overwriting it – there’s no need to
throw away anything we might already have.

Note that we shouldn’t unset() the keys, because initShallowFallback()
takes references to them. Overwrite them with empty arrays instead.

Bug: T342418
Change-Id: Ib1c620d4f21cc1dba478a837271778dae1000d52