Make "pluralRules" caches static in LocalisationCache
The source for this piece of code is a hard-coded set of two files,
hard-coded in LocalisationCache::getPluralFiles():
- languages/data/plurals.xml
- languages/data/plurals-mediawiki.xml

These files don't change between tests. There should never be a
reason to parse them a second time. The in-memory caching (that
already exists) can as well be static.

While this does not have a huge impact on test runtime, the effect
can still be seen. In a very expensive test like
LanguageConverterFactoryTest this method was called many 100 times.
Now it's called a single time. Runtime of this test goes down from
6.8s to 6.1s.

Bug: T225730
Change-Id: Id77d0f5034c4ce4a054d00ada0d12087066ae816