rdbms: Drop srvCache from Database
It's only used in-class for getMySqlServerVariant() and publicly in
IDatabase::getServerVersion() and IDatabase::getServerInfo().

The function is not called often (not called once during the majority
of requests to any web entry point), and the SQL command is quite fast.

I wanted to put the cache in front of a caller (as it's the
responsibility of callers to cache database queries) but not one place
really needs it. It's basically Special:Version and Siteinfo API.

Originally added with If667f761b2173226ccf1129e in 2019.

Bug: T342564
Change-Id: I45e0918fce0b751e2d129a0c64d5755e2314c65e