changetags: Use LogPage::DELETED_RESTRICTED in ChangeTagsLogItem
ChangeTagsLogItem::canView is used to detect if the log entry can be
shown at all to the user, this is the case for all users when the log
entry does not contains suppressed parts. When there are suppressed
parts the user must have permissions for it to see the log entry. It
needs just one suppressed part like user name, not all as the constant
indicates.
The check for all is not necessary and the called function does not work
like that.
This reverts the change to this file from 8a26fa05

Change-Id: Ia9cc015315ede4bcb96095a2f439842cf7102f72