logging: Clarify meaning of `userToolLinks($edits)` parameter
== Optional ==

It was described as "optional" but it's not a "typical" optional
parameter because its presence has non-trivial consequences on the
output in a way that can't be ignored by callers.

It sounded like passing it would be faster for cases where the caller
has computed it already and can pass it to allow re-use, and
otherwise the function will compute it on-demand, but behave the same
way logically. Except, that isn't true at all.

== Null fallback ==

Replace the overcomplicated `intval( $edits ) === 0 && $edits !== 0`
expression with an explicit `=== null` check. The old logic actually
also accepted values that cast to zero, but are not zero.

This originates from a misunderstanding in a 2012 code review,
where it was suggested to use `(int)` on the non-null before
returning it. But this author did not do that, and instead created the
above complicated conditional. [1]

Change 26457, PS2:
> ```
> $count = !is_null( $edits ) ? $edits : $user->getEditCount();
> ```
> Should `? $edits` be `? intval( $edits )`?

Change 26457, PS3:
```
if ( intval( $edits ) === 0 && $edits !== 0 ) {
  $edits = $user->getEditCount();
}
```

I believe the intention of the reviewer was:

```
$count = $edits !== null ? (int)$edits : $user->getEditCount();
```

[1] https://gerrit.wikimedia.org/r/c/mediawiki/core/+/26457/2

== Redundant null ==

Follows-up I62faf9d042a92. Remove another redundant ternary from
a UserEditTracker caller, in LogFormatter, as getEditCount() already
returns `int|null` based on User->getId().

Change-Id: Iaf68ec373f2458554cd028844e6531244cd28356