context: Simplify RequestContext::getSkin() logic
I had to read this 5x to convince myself that it wouldn't discard
the assigned `$this->skin` object (from the elseif branch in the middle)
in the `if ( $skinName !== null )` branch below.

It appears to work fine, but it may be clearer to avoid the indirection
through checking that $skinName got assigned and rather move those
branches together.

This coincidentally matches the evaluation order of fetchSkinName().

Change-Id: Ia20136b8f04c8e8d89d29b97cdd99cb444c09838