Record Parsoid version in extension data to allow rollback if necessary
This allows any bad cached parses due to a train deploy to be selectively
rolled back in the RejectParserCacheValue hook, which provides some
operational insurance against corrupted caches.  The version is also
added to the debug information in the HTML footer to aid diagnosis
of any issue in real time.

Depends-On: I3d3caabd959c1ba16f4dc702c2eae38d5d4dcb14
Change-Id: Ibb37a82ec0ce764aefd8c9fab2868073a66301ec