Make minimumGroupingDigits an integer and add documentation
All Messages*.php variables are supposed to be documented in
MessagesEn.php.

I don't see the point in having Language::minimumGroupingDigits() return
null, and then expecting every caller to cast null to zero.

The default for English is properly 1, not 0, since 3-digit numbers do
not need grouping. We don't write 999 as ",999". Setting it to 1
instead of null should improve the performance of the JS commafyNumber()
by avoiding the unnecessary loop body. Change the QUnit test data to
confirm that this works.

In the PHP implementation, adjust the comparison to avoid the
preg_match() when minimumGroupingDigits is 1.

Change-Id: I6679d69e7094502303389039550e17a47189e2dc