tests: Unbreak and actually run ExportDemoTest
This test was not being run because the docs directory wasn't included
in the PHPUnit XML config files. As a direct consequence, the test was
extremely broken: it used the non-namespaced DumpTestCase class which no
longer exists, it referenced XML files with the wrong path, and then the
test had other non-error failures to fix.

Change-Id: I188cbb1de68d39c7a484d4c6f808678af80d11ab