ResourceLoader: use Function constructor for module implementations
Profiling in Chromium 115 of a VE request with a mw.loader.store cache
hit shows that the largest domEval() takes 59ms, mostly DOM-related
overhead. Using the Function constructor reduces the time for that call
to 30ms.

Profiling in Firefox 116 shows that it is 41ms either way, there is no
significant DOM overhead.

So, use the Function constructor instead of DOM eval for normal RL
module implementations.

User scripts rely on e.g.

function foo() {}

creating window.foo, which is not the case for the Function constructor.
So for them, continue to use domEval().

The Function constructor propagates exceptions, which is beneficial,
since domEval() was called in a try/catch block. Now the catch should
be reachable.

Change-Id: I51b8535b2b21a3ed62e05c727a06dee09ba0a7b7