core - client side preferences inline script
Changes:
- Add/Modify client preference classes to documentElement,
- Only uses existing classes, for maintainability. For new feature,
the new class must be added a couple of weeks before the feature is deployed,
so that the feature class is already in the cache when the feature is deployed
- Add mw.user.clientPrefs as an API to manage Preferences. This exposes set and get methods
- When set method is called it saves to a cookie which saves key value pairs of
feature and its value. The value corresponds with a suffix on the class.
- get method always uses HTML to obtain result.
- Tests are defined for set and get methods to ensure that the API does not have unexpected
side effects and caters for the case where two tabs are open and the cookie may not reflect
the classes on the HTML element
- A prefix "-clientpref-" is required on any class managed by this capability.
The purpose of this is to aid discoverability and to act as an allow list for which classes
can be modified by the API
- Using '~' instead of '=' and '!' instead of ',' for separators as they require no url encoding
and do not interfere with the delimiters and special characters
- This is a breaking change for Vector 2022's existing limited width feature. For this
reason this patch should be merged around the same time as the Vector 2022 patch to ensure
backwards compatibility.
- No backwards compatibility is provided by core given the previous capability was Vector
specific. Backwards compatibility will be handled in the skin. Please see:
I120f8f7114b33d2cfbd1c3c57ebf41f8b2d7fec4

Note:
This patch doesn't provide tests for the inline script because
* It is not possible to add this to this patch at the current time since we need to determine
whether or not it is okay to call file_get_contents at such a critical time
* the tests are written in Jest due to limitations with our QUnit test framework.

For development purposes a standalone test patch was added I6ad496bb25a4bd523fa69f2e6d936d7eb643793e
A task will be created on merge to explore incorporating these tests in the codebase.

Bug: T339268
Bug: T341720
Change-Id: I1e635f843ac9b2f248b1f7618134598e80291b38