rdbms: Roll up once-used small private/protected methods to the caller
We have done a similar thing with Database class and cleaned up a lot.
Now doing it for LB/LBF and their subclasses.

The only exception was one-liner that I repeated it (it was called
twice) as it wasn't de-duplicating any actual logic. Assuming if it's
not master, it's replica is baked into most of rdbms.

Bug: T326274
Change-Id: Ic2498c67f523612af18cbfb7204ee40a01fd79de