ResourceLoader: Simplify module serialization
When a module is saved to the localStorage cache, instead of piecing
the module back together from the parameters passed to
mw.loader.implement(), cast the declarator function to a string.

Simplify mw.inspect.getModuleSize() and its test cases.

Keep mw.loader.implement() for b/c, but do not cache modules created by
it.

Bug: T343407
Change-Id: I698380a27d119aa901d4ecee93dde62f8c2044ee