LocalisationCache: Correctly merge preload data
This fixes a bug in change Ib1c620d4f2 (commit 7355d75c5f) which only
becomes apparent when change I7ec2d87c0f (commit cf8e22e1e4) is also
applied. In this scenario, it is possible for the core-only data to be
loaded first, populating $this->data with an incomplete version of
namespaceNames. Then, the preload data is merged into $this->data[$code]
with the += operator, which discards the preload’s namespaceNames
because $this->data[$code] already contains namespaceNames. And finally,
all the keys in the preload data are marked as loaded, so the incomplete
core-only namespace names are now incorrectly marked as fully loaded,
losing all the extension namespace names. The same goes for dateFormats
as well (see PRELOADED_KEYS).

To fix this, properly merge the $preload data using $this->mergeItem().
(I think another option would be to have the preload data override all
the $this->data[$code] keys that aren’t marked as loaded yet.)

Bug: T342418
Bug: T343343
Bug: T343375
Change-Id: Icc3324aca7291874259925e5f7e66f28c0555ecb