Add "mediawiki.page.teleportTarget" module to core
Introduces a new ResourceLoader module, "mediawiki.page.teleportTarget".
This script creates a new empty div with an ID and appends it to
the end of the document body. An accompanying stylesheet has also
been added that sets this element's z-index to that of the
@z-index-overlay token. Skins can override this z-index value if
necessary.

This is done in order to provide a stable target that client-side
UI code can use if a dialog or similar modal element needs to be
teleported without causing collisions with other elements on the page.

This module has been added as a dependency of the MW Vue module.
The custom Vue.createMWApp method now provides the appended DOM
element under the "CdxTeleportTarget" key so that components like
Codex Dialog will use it automatically.

MW Skins should consider applying skinStyles to this element
to ensure that it's styles match those of the page body text elements.

Bug: T343476
Change-Id: Ia79b76c0e73890dd18477e5c3ea307d753cedb3a