Auth: Introduce several AuthManager session keys as constants
This patch introduces session keys as constants as they repeat across
this class but are the same strings. Making them as constants will ensure
they're the same across usage and avoid errors.

If we ever want to change this key in the future, we can just change the
value of the constant and no need to change all occurrences in the entire
file.

== Introduces ==
+ AUTHN_STATE
+ ACCOUNT_CREATION_STATE
+ ACCOUNT_LINK_STATE
+ AUTOCREATE_BLOCKLIST

NOTE: Constants created as public but marked as internal to be used by
relevant tests in core.

Change-Id: I49e4ea01f0f8b548e32f88653e67ba4058c8bc07